refactor: extract typing state cleanup
