+++ src/views/ChatView.vue
@@ -68,6 +68,7 @@ import { presetState, getEffectivePreset, getEffectivePresetId } from '@/core/st
 import { useVirtualScroll } from '@/composables/chat/useVirtualScroll.js';
 import { useGenerationRegistry } from '@/composables/chat/useGenerationRegistry.js';
 import { createPromptMetadataSnapshots } from '@/composables/chat/usePromptMetadataSnapshots.js';
+import { useTypingStateCleanup } from '@/composables/chat/useTypingStateCleanup.js';
 import { useSidebarResizer } from '@/composables/ui/useSidebarResizer.js';
 import { sendMessageNotification, clearMessageNotifications, startGenerationNotification, stopGenerationNotification } from '@/core/services/notificationService.js';
 import { addNotification } from '@/core/states/notificationsState.js';
@@ -187,6 +188,7 @@ const {
     markGenerationPersisted,
     clearPersistedGeneration
 } = useGenerationRegistry();
+const { clearTypingStateForMessage } = useTypingStateCleanup({ currentMessages, getChatData, db });
 
 // --- Component State ---
 const chatViewRoot = ref(null);
@@ -3258,26 +3260,6 @@ function startGeneration(char, text, existingMsgIndex = -1, onAbort = null, guid
         }
     }, 100);
 
-    const clearTypingStateForMessage = async () => {
-        const idx = currentMessages.value.findIndex(m => m.id === msgId);
-        if (idx !== -1) {
-            currentMessages.value[idx].isTyping = false;
-        }
-
-        try {
-            const data = await getChatData(char.id);
-            if (data && data.sessions[sessionId]) {
-                const dbIdx = data.sessions[sessionId].findIndex(m => m.id === msgId);
-                if (dbIdx !== -1) {
-                    data.sessions[sessionId][dbIdx].isTyping = false;
-                    await db.saveChat(char.id, data);
-                }
-            }
-        } catch (dbErr) {
-            console.error('[generation] Failed to clear typing state:', dbErr);
-        }
-    };
-
     const restoreState = async (isError = false) => {
         if (_bgUpdateTimer) { clearTimeout(_bgUpdateTimer); _bgUpdateTimer = null; }
         if (getGenerationState(char.id)?.timerId) clearInterval(getGenerationState(char.id).timerId);
@@ -3363,23 +3345,12 @@ function startGeneration(char, text, existingMsgIndex = -1, onAbort = null, guid
 
         // CRITICAL: Ensure isTyping is ALWAYS cleared, even if DB operations fail
         const ensureTypingCleared = async () => {
-            const idx = currentMessages.value.findIndex(m => m.id === msgId);
-            if (idx !== -1) {
-                currentMessages.value[idx].isTyping = false;
-            }
-            // Also clear in DB to prevent phantom typing on reload
-            try {
-                const data = await getChatData(char.id);
-                if (data && data.sessions[sessionId]) {
-                    const dbIdx = data.sessions[sessionId].findIndex(m => m.id === msgId);
-                    if (dbIdx !== -1) {
-                        data.sessions[sessionId][dbIdx].isTyping = false;
-                        await db.saveChat(char.id, data);
-                    }
-                }
-            } catch (dbErr) {
-                console.error('[onError] Failed to clear isTyping in DB:', dbErr);
-            }
+            await clearTypingStateForMessage({
+                charId: char.id,
+                sessionId,
+                msgId,
+                errorLabel: '[onError]'
+            });
         };
 
         try {
@@ -3575,7 +3546,7 @@ function startGeneration(char, text, existingMsgIndex = -1, onAbort = null, guid
             if (_bgUpdateTimer) { clearTimeout(_bgUpdateTimer); _bgUpdateTimer = null; }
 
             if (!currentState || currentState.genId !== genId) {
-                await clearTypingStateForMessage();
+                await clearTypingStateForMessage({ charId: char.id, sessionId, msgId });
                 ensureStaleCleanup();
                 window.dispatchEvent(new CustomEvent('chat-generation-ended', { detail: { charId: char.id, sessionId: sessionId } }));
                 return;
@@ -3587,7 +3558,7 @@ function startGeneration(char, text, existingMsgIndex = -1, onAbort = null, guid
             // Guard against race with abort
             const hasCompletionPayload = !!(response || finalReasoning || meta?.partialError);
             if (controller.signal.aborted && !hasCompletionPayload) {
-                await clearTypingStateForMessage();
+                await clearTypingStateForMessage({ charId: char.id, sessionId, msgId });
                 ensureCleanup();
                 window.dispatchEvent(new CustomEvent('chat-generation-ended', { detail: { charId: char.id, sessionId: sessionId } }));
                 return;
@@ -3764,23 +3735,12 @@ function startGeneration(char, text, existingMsgIndex = -1, onAbort = null, guid
             console.error('[onComplete] Completion handler failed:', completeErr);
             // Ensure cleanup and isTyping cleared even if handler fails
             ensureCleanup();
-            // Clear isTyping in both reactive state and DB
-            const idx = currentMessages.value.findIndex(m => m.id === msgId);
-            if (idx !== -1) {
-                currentMessages.value[idx].isTyping = false;
-            }
-            try {
-                const data = await getChatData(char.id);
-                if (data && data.sessions[sessionId]) {
-                    const dbIdx = data.sessions[sessionId].findIndex(m => m.id === msgId);
-                    if (dbIdx !== -1) {
-                        data.sessions[sessionId][dbIdx].isTyping = false;
-                        await db.saveChat(char.id, data);
-                    }
-                }
-            } catch (dbErr) {
-                console.error('[onComplete] Failed to clear isTyping in DB:', dbErr);
-            }
+            await clearTypingStateForMessage({
+                charId: char.id,
+                sessionId,
+                msgId,
+                errorLabel: '[onComplete]'
+            });
             window.dispatchEvent(new CustomEvent('chat-generation-ended', { detail: { charId: char.id, sessionId: sessionId } }));
         }
             },
